requirements.txt
numpy==1.24.3
matplotlib==3.7.2
Pillow==10.0.0
Flask==2.3.3
